Move module manifest properties to 'update' step & add new items
For some unexplained reason, the update-manifest command was removing the NestedModules property. To fix, I only use the New-Manifest command to create the file and use the Update-Manifest command to add all the manifest properties. This seems to retain all information. I also added a few new properites, such as copyright and
Help URI.